backend/requirements.txt
# Web Framework
fastapi==0.118.0
uvicorn[standard]==0.24.0

# Environment variables
python-dotenv==1.0.0

# Computer Vision and AI
opencv-python-headless==4.8.1.78
mediapipe==0.10.14
ultralytics==8.0.206
numpy>=1.26.0,<1.27.0

# Data Processing and Visualization
pandas==2.3.3
matplotlib==3.10.6
fpdf2==2.8.4

# HTTP Client
httpx==0.28.1

# AI and Language Models
google-generativeai==0.3.2

# Environment Management
python-dotenv==1.0.0

# WebSocket Support (included with FastAPI/uvicorn)
websockets==15.0.1

# Additional utilities
pydantic==2.11.10
python-multipart==0.0.20